+++ src/main/java/com/alexsid/lesson15/pets/Pet.java
@@ -1,7 +1,6 @@
-package com.alexsid.lesson1517.pets;
+package com.alexsid.lesson15.pets;
 
-import com.alexsid.lesson1517.Person;
-import com.alexsid.lesson1517.Person;
+import com.alexsid.lesson15.Person;
 
 public interface Pet {
     int getId();
