+++ src/main/java/com/alexsid/lesson24/ControllerHi.java
@@ -1,0 +1,4 @@
+package com.alexsid.lesson24;
+
+public class ControllerHi {
+}
