+++ src/main/java/com/alexsid/lesson24/ModelHi.java
@@ -1,0 +1,4 @@
+package com.alexsid.lesson24;
+
+public class ModelHi {
+}
